# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SteamShare API - Steam account marketplace with points, XP, and badges
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [users]
      summary: Top users by XP level
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user's public profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /users/{userId}/accounts:
    get:
      operationId: getUserAccounts
      tags: [users]
      summary: Get all accounts posted by a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"

  # Ad Links
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserPublic:
      type: object
      required: [id, username, xp, level, points, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        points:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        badgeName:
          type: ["string", "null"]
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required:
        [
          id,
          username,
          xp,
          level,
          points,
          totalAccounts,
          totalLikesReceived,
          createdAt,
        ]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        points:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        badgeName:
          type: ["string", "null"]
        totalAccounts:
          type: integer
        totalLikesReceived:
          type: integer
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Account:
      type: object
      required:
        [
          id,
          userId,
          username,
          title,
          description,
          games,
          pointsCost,
          isAvailable,
          likesCount,
          claimsCount,
          createdAt,
        ]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        posterUsername:
          type: string
        posterAvatarUrl:
          type: ["string", "null"]
        posterNameColor:
          type: ["string", "null"]
        posterBadgeType:
          type: ["string", "null"]
        posterBadgeIconUrl:
          type: ["string", "null"]
        posterBadgeIconLink:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        games:
          type: array
          items:
            type: string
        pointsCost:
          type: integer
        isAvailable:
          type: boolean
        likesCount:
          type: integer
        claimsCount:
          type: integer
        viewCount:
          type: integer
        workingVotes:
          type: integer
        notWorkingVotes:
          type: integer
        unlockMethod:
          type: string
          enum: [login, like, comment, vip]
        userHasLiked:
          type: boolean
        userHasCommented:
          type: boolean
        createdAt:
          type: string
          format: date-time
...