# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SteamShare API - Steam account marketplace with points, XP, and badges
 * OpenAPI spec version: 0.1.0
 */
import type { AccountUnlockMethod } from './accountUnlockMethod';

export interface Account {
  id: number;
  userId: number;
  username: string;
  posterUsername?: string;
  /** @nullable */
  posterAvatarUrl?: string | null;
  /** @nullable */
  posterNameColor?: string | null;
  /** @nullable */
  posterBadgeType?: string | null;
  /** @nullable */
  posterBadgeIconUrl?: string | null;
  /** @nullable */
  posterBadgeIconLink?: string | null;
  title: string;
  description: string;
  games: string[];
  pointsCost: number;
  isAvailable: boolean;
  likesCount: number;
  claimsCount: number;
  viewCount?: number;
  workingVotes?: number;
  notWorkingVotes?: number;
  unlockMethod?: AccountUnlockMethod;
  userHasLiked?: boolean;
  userHasCommented?: boolean;
  createdAt: Date;
}